+++ cli.ts
@@ -1,4 +1,5 @@
 import {
+	checkVersion,
 	Command,
 	deploy,
 	init,
@@ -8,10 +9,11 @@ import {
 	signup,
 	start,
 	upgrade,
+	util,
 } from "./cli/mod.ts";
 import "https://deno.land/std@0.179.0/dotenv/load.ts";
 
-export const version = "v0.3.6";
+export const version = "v0.3.7";
 
 const command = new Command()
 	.name("arkiver")
@@ -23,7 +25,10 @@ command
 	.command("login", "Login to RoboArkiver")
 	.option("-e, --email <email:string>", "Email address")
 	.option("-p, --password <password:string>", "Password")
-	.action(login.action);
+	.action(async (opts, ...args) => {
+		await checkVersion(version);
+		await login.action(opts, ...args);
+	});
 
 // signup
 command
@@ -31,7 +36,10 @@ command
 	.option("-e, --email <email:string>", "Email address")
 	.option("-p, --password <password:string>", "Password")
 	.option("-u, --username <username:string>", "Username")
-	.action(signup.action);
+	.action(async (opts, ...args) => {
+		await checkVersion(version);
+		await signup.action(opts, ...args);
+	});
 
 // signout
 command.command("logout", "Logout from RoboArkiver").action(logout.action);
@@ -42,13 +50,19 @@ command
 	.option("--public", "Make arkive public")
 	.option("--major", "Deploy as major version")
 	.arguments("<dir:string>")
-	.action(deploy.action);
+	.action(async (opts, ...args) => {
+		await checkVersion(version);
+		await deploy.action(opts, ...args);
+	});
 
 // delete
 command
 	.command("delete", "Delete arkive")
 	.arguments("<id:number>")
-	.action(async (_, id) => await remove.action(id));
+	.action(async (_, id) => {
+		await checkVersion(version);
+		await remove.action(id);
+	});
 
 // start
 command
@@ -65,15 +79,25 @@ command
 		collect: true,
 	})
 	.option("--no-gql", "Disable GraphQL server")
-	.action(start.action);
+	.option("--no-db", "Don't connect to MongoDB")
+	.action(async (opts, ...args) => {
+		util.logHeader(version);
+		await checkVersion(version);
+		await start.action(opts, ...args);
+	});
 
-//init
+// init
 command
 	.command("init", "Initialize a new arkive project")
 	.arguments("<dir:string>")
 	.option("--overwrite", "Overwrite existing files")
-	.action(init.action);
+	.action(async (opts, ...args) => {
+		util.logHeader(version);
+		await checkVersion(version);
+		await init.action(opts, ...args);
+	});
 
+// upgrade
 command
 	.command("upgrade", "Upgrade arkiver to latest version")
 	.action(async () => await upgrade.action(version));