+++ examples/simple/transferHandler.ts
@@ -1,4 +1,4 @@
-import { EventHandlerFor, formatUnits, logger } from "./deps.ts";
+import { EventHandlerFor, formatUnits } from "./deps.ts";
 import erc20 from "./erc20.ts";
 import { Balance } from "./entities.ts";
 
@@ -10,8 +10,7 @@ export const transferHandler: EventHandlerFor<typeof erc20, "Transfer"> =
 
 		const address = event.address;
 
-		// store.retrieve() is a wrapper around Map.get() that will
-		// call the provided function if the key is not found in the store.
+		// store.retrieve() will return the value if it exists in the store, otherwise it will run the function and store the result
 		const decimals = await store.retrieve(
 			`${address}:decimals`,
 			async () =>
@@ -20,7 +19,15 @@ export const transferHandler: EventHandlerFor<typeof erc20, "Transfer"> =
 					functionName: "decimals",
 					address,
 				}),
-			{ ttl: 1000 * 60 * 60 * 24 * 7 },
+		);
+
+		// reduce rpc calls in case you have multiple events in the same block
+		const timestamp = await store.retrieve(
+			`${event.blockHash}:timestamp`,
+			async () => {
+				const block = await client.getBlock({ blockHash: event.blockHash });
+				return Number(block.timestamp);
+			},
 		);
 
 		const parsedValue = parseFloat(formatUnits(value, decimals));
@@ -29,35 +36,39 @@ export const transferHandler: EventHandlerFor<typeof erc20, "Transfer"> =
 			await store.retrieve(
 				`${from}:${address}:balance`,
 				async () => {
-					return await Balance.findOne({ account: from }) ??
-						new Balance({
-							amount: 0,
-							token: address,
-							account: from,
-						});
+					const balance = await Balance
+						.find({ account: from, token: address })
+						.sort({ timestamp: -1 })
+						.limit(1);
+					return balance[0]?.amount ?? 0;
 				},
 			),
 			await store.retrieve(
 				`${to}:${address}:balance`,
-				async () =>
-					await Balance.findOne({ account: to }) ??
-						new Balance({
-							amount: 0,
-							token: address,
-							account: to,
-						}),
+				async () => {
+					const balance = await Balance
+						.find({ account: from, token: address })
+						.sort({ timestamp: -1 })
+						.limit(1);
+					return balance[0]?.amount ?? 0;
+				},
 			),
 		]);
 
-		senderBalance.amount -= parsedValue;
-		receiverBalance.amount += parsedValue;
+		const senderNewBalance = senderBalance - parsedValue;
+		const receiverNewBalance = receiverBalance + parsedValue;
 
-		store.set(
-			`${from}:${address}:balance`,
-			senderBalance.save().catch((e) => logger.error(e)),
-		);
-		store.set(
-			`${to}:${address}:balance`,
-			receiverBalance.save().catch((e) => logger.error(e)),
-		);
+		// save the new balances to the database
+		Balance.create({
+			account: from,
+			amount: senderNewBalance,
+			token: address,
+			timestamp,
+		});
+		Balance.create({
+			account: to,
+			amount: receiverNewBalance,
+			token: address,
+			timestamp,
+		});
 	};